backend/requirements.txt
asgiref==3.8.1
channels==4.1.0
colorama==0.4.6
Django==5.0.8
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-shortuuidfield==0.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docx2pdf==0.1.8
lxml==5.2.2
pillow==10.4.0
PyJWT==2.9.0
python-docx==1.1.2
# pywin32==306
rest-framework-simplejwt==0.0.2
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
psycopg2==2.9.10
hypercorn==0.17.3
django-notifications-hq==1.8.3